requirements.txt
pycryptodome
requests
base64
time
subprocess
enum